// src/app/interfaces/FichaConectados.ts
export interface FichaConectado {
    Tribo: number;
    Lider: string;
    Cep: string;
    Endereco: string;
    Nome: string;
    Sexo: number;
    EstadoCivil: number;
    Nascimento: Date;
    Telefone: string;
    Email: string;
    ContatoEmergencial: string;
    Crianca: boolean;
    Cuidados: boolean;
    Idade: number;
    DescricaoCuidados: string;
    Siao: number;
}